requirements.txt
mysql-connector-python
Pillow
tk 